Time: 3 ms (47.84%), Space: 48.5 MB (61.52%) - LeetHub

+++ 0713-subarray-product-less-than-k/0713-subarray-product-less-than-k.java
@@ -3,7 +3,7 @@ class Solution {
         int low=0;
         int high=0;
 
-        long product=1;
+        int product=1;
         int count=0;
         while(high<nums.length)
         {
